refactor(discord): decompose inflight.rs into child modules (#3835)
Pure verbatim move (zero behavior change). Relocates two cohesive clusters
out of the oversized `inflight.rs` hot-state parent:

- typed status-panel / current-message ownership writes -> new
  `inflight/ownership_ops.rs`
- staleness predicates + orphan-lock / rebind-origin reap helpers ->
  the existing `inflight/rebind_reap.rs`

Every externally-referenced symbol is re-exported from the `inflight.rs`
facade at its original visibility, so no caller outside the inflight module
changes its import path (`inflight::*` paths stay byte-identical). The watcher
progress / terminal-commit / relay-watermark cluster (also #3835) had already
moved to `inflight/watcher_state.rs` in a prior PR.

Production LoC: inflight.rs 3077 -> 2275; giant baseline lowered to lock the win.

Invariants preserved (moved critical sections verbatim, not rewritten):
- lock-atomic offset/owner-field writes (whole flock-held section moved intact)
- identity-guarded saves/clears still reject stale-row writes
- offset-monotonicity signal severity + skip behavior unchanged
- clear_inflight_by_tmux_name stays the canonical tmux-keyed cleanup (in core)
