+++ ARCHITECTURE.md
@@ -453,6 +453,7 @@ src/
 │   │   │   ├── finalizer_identity.rs
 │   │   │   ├── model.rs
 │   │   │   ├── orphan_relay_reclaim.rs
+│   │   │   ├── ownership_ops.rs
 │   │   │   ├── rebind_reap.rs
 │   │   │   ├── store.rs
 │   │   │   └── watcher_state.rs
