+++ src/services/discord/inflight.rs
@@ -30,17 +30,35 @@ use store::inflight_provider_dir;
 pub(in crate::services::discord::inflight) use store::inflight_state_path;
 pub(crate) use store::lock_inflight_state_path;
 
-// #3715: move the rebind-origin dead-watcher/orphan-lock helpers into a capped
-// sibling module so this hot state parent stays below the frozen production-LoC
+// #3715 / #3835: the rebind-origin dead-watcher/orphan-lock helpers PLUS the
+// staleness predicates and orphan-lock / rebind-origin reap helpers live in this
+// capped sibling so this hot state parent stays below the frozen production-LoC
 // baseline without changing call-site names.
 mod rebind_reap;
-use self::rebind_reap::{
-    RuntimeWatcherLiveness, WatcherLiveness, dead_watcher_rebind_structurally_reapable,
-    inflight_json_path_for_lock, is_inflight_json_lock_path, metadata_mtime_unix_secs,
-    rebind_origin_age_secs, watcher_runtime_activity_recent,
+// Facade re-exports: keep every reap/staleness symbol still referenced by
+// discord-module / inflight-core lib code at its original `pub(super)` visibility
+// so `inflight::*` paths stay byte-identical after the #3835 move.
+pub(super) use self::rebind_reap::{
+    INFLIGHT_STALENESS_THRESHOLD_SECS, RebindReapOutcome, emit_reap_abandoned_rebind_origin,
+    inflight_state_is_stale, ownerless_external_input_inflight_is_stale, parse_started_at_unix,
+    parse_updated_at_unix, reap_abandoned_rebind_origin_locked, reap_orphan_inflight_locks,
+    rebind_origin_deadline_secs_env, should_reap_abandoned_rebind_origin,
+    sweep_reap_dead_watcher_rebind_origin,
 };
+// Parent-internal helper (private in the parent before the move): re-imported so
+// inflight-core resolves it unchanged, WITHOUT widening it to the discord module.
+use self::rebind_reap::{reap_abandoned_rebind_origin_locked_in_root, rebind_origin_age_secs};
+// Symbols only referenced by the in-file `#[cfg(test)]` modules (their lib callers
+// moved into `rebind_reap`): re-imported test-only so the test modules resolve
+// `super::*` unchanged without emitting unused-import warnings in the lib build.
 #[cfg(test)]
-use self::rebind_reap::{proven_dead_from_signals, should_reap_dead_watcher_rebind_origin};
+use self::rebind_reap::{
+    DEAD_WATCHER_PROVEN_DEAD_SECS, ORPHAN_LOCK_REAP_MIN_AGE_SECS,
+    REBIND_ORIGIN_DEADLINE_SECS_DEFAULT, WatcherLiveness,
+    ownerless_external_input_inflight_is_stale_at, proven_dead_from_signals,
+    reap_dead_watcher_rebind_origin_locked, reap_dead_watcher_rebind_origin_locked_in_root,
+    reap_orphan_inflight_locks_in_root, should_reap_dead_watcher_rebind_origin,
+};
 
 mod watcher_state;
 pub(in crate::services::discord) use self::watcher_state::{
@@ -53,6 +71,20 @@ pub(in crate::services::discord) use self::watcher_state::{
 use self::watcher_state::{
     commit_watcher_terminal_delivery_locked_in_root,
     persist_watcher_relay_watermark_locked_in_root, persist_watcher_stream_progress_locked_in_root,
+};
+
+// #3835: typed status-panel / current-message ownership writes (the #3077
+// panel-lifecycle race family) moved into a child module; re-exported below so
+// `inflight::*` paths stay byte-identical for discord-module callers.
+mod ownership_ops;
+pub(in crate::services::discord) use self::ownership_ops::{
+    StatusPanelBindGuard, StatusPanelBindOutcome, StatusPanelClearGuard, bind_status_panel,
+    clear_current_msg_if_matches, clear_status_panel_if_current,
+};
+#[cfg(test)]
+use self::ownership_ops::{
+    bind_status_panel_in_root, clear_current_msg_if_matches_in_root,
+    clear_status_panel_if_current_in_root,
 };
 
 // #3960: producer-liveness TOCTOU reclaim for orphaned `SessionBoundRelay`
@@ -108,74 +140,6 @@ const HOT_SWAP_HANDOFF_MAX_AGE_SECS: u64 = 900; // 15 minutes
 /// "3 consecutive restarts" budget, not a per-process retry cap.
 pub(super) const RECOVERY_RELAY_RESTART_ATTEMPT_BUDGET: u32 = 3;
 
-/// #1446 stall-deadlock recovery: an inflight state is treated as "stale"
-/// (i.e. the dispatch that wrote it almost certainly already terminated
-/// without cleanup) when its persisted `updated_at` has not advanced for
-/// this many seconds. THREAD-GUARD uses this exact threshold; the
-/// stall-watchdog uses `2x` to stay strictly more conservative than any
-/// caller that has already observed the state directly.
-///
-/// `updated_at` is rewritten on every `save_inflight_state` call but is
-/// **not** a true heartbeat — a healthy foreground model/tool call can
-/// legitimately go silent for multiple minutes (long Bash, slow LLM
-/// stream, large Read).
-///
-/// History: aligned with `placeholder_sweeper::ABANDON_THRESHOLD_SECS` (then
-/// 300s) until #2427 (#2436 / #2437 / #2438) added explicit-signal wires (pane
-/// death, heartbeat-gap sweeper, generation-mismatch invalidate, TurnCompleted
-/// guard), relaxing the sweeper to a pure safety net (abandon timer 1800s). 300s
-/// is retained here because it gates **new**-dispatch THREAD-GUARD + the
-/// stall-watchdog (#1446): both want fast recovery once an explicit signal failed
-/// to fire (a false-positive cleanup of a live turn is far worse than delay).
-pub(super) const INFLIGHT_STALENESS_THRESHOLD_SECS: u64 = 300;
-
-/// #3581: default deadline (seconds) after which an unadopted, never-progressed
-/// `rebind_origin` row is reaped. Far shorter than the placeholder sweeper's
-/// 1800s safety net so the wedge-causing orphan drains within a handful of
-/// sweep ticks, but comfortably longer than the ~8s TUI-direct adoption
-/// backstop window so a legitimate `/api/inflight/rebind` → TUI-adopt handoff
-/// is never raced. Overridable via `AGENTDESK_REBIND_ORIGIN_DEADLINE_SECS`.
-pub(super) const REBIND_ORIGIN_DEADLINE_SECS_DEFAULT: u64 = 120;
-
-/// #3641: orphan sidecar lock files are cosmetic after process death because
-/// `flock` releases with the fd, but stale paths can accumulate forever once the
-/// matching `.json` state row has been cleaned/quarantined. Keep the reap
-/// conservative so a fresh lock created just before its `.json` row is written is
-/// never touched.
-pub(super) const ORPHAN_LOCK_REAP_MIN_AGE_SECS: i64 = 3600;
-
-/// #3581: floor for the env-overridden rebind-origin deadline. Guards against a
-/// pathologically small (or zero) override reaping rows before the adoption
-/// backstop window can complete.
-const REBIND_ORIGIN_DEADLINE_SECS_MIN: u64 = 30;
-
-/// #3635: minimum quiescence (seconds) of *all* runtime liveness signals before
-/// a Watcher-owned rebind-origin row is treated as "proven dead" and made
-/// eligible for the dead-watcher reap path. Deliberately far larger than the
-/// stall-watchdog's positive-liveness window
-/// (`STALL_WATCHDOG_POSITIVE_LIVENESS_SECS` = 120s): a live Watcher that is
-/// merely between turns (mailbox idle, no fresh jsonl byte) must NEVER be
-/// false-classified as dead and reaped, because #3154/#3540 require a live
-/// Watcher rebind to survive so it can re-adopt the session across a restart.
-/// A 10-minute conservative floor means a false-negative (a genuinely dead
-/// watcher whose reap is merely delayed) is the only failure mode — never a
-/// false-positive (a live watcher reaped). The orphan is harmless while it
-/// lingers (#3631/PR #3634 classifies it idle), so delay is acceptable.
-pub(super) const DEAD_WATCHER_PROVEN_DEAD_SECS: u64 = 600;
-
-/// #3581: resolve the rebind-origin reap deadline. Reads
-/// `AGENTDESK_REBIND_ORIGIN_DEADLINE_SECS`; on absence / parse failure falls
-/// back to [`REBIND_ORIGIN_DEADLINE_SECS_DEFAULT`]. Any explicit value is
-/// clamped up to [`REBIND_ORIGIN_DEADLINE_SECS_MIN`] so an operator cannot
-/// accidentally configure a reap that races the adoption backstop.
-pub(super) fn rebind_origin_deadline_secs_env() -> u64 {
-    std::env::var("AGENTDESK_REBIND_ORIGIN_DEADLINE_SECS")
-        .ok()
-        .and_then(|raw| raw.trim().parse::<u64>().ok())
-        .map(|secs| secs.max(REBIND_ORIGIN_DEADLINE_SECS_MIN))
-        .unwrap_or(REBIND_ORIGIN_DEADLINE_SECS_DEFAULT)
-}
-
 /// #3581: current Unix epoch seconds (wall clock). Used to stamp a
 /// rebind-origin row's birth time at creation.
 pub(super) fn now_unix() -> i64 {
@@ -185,179 +149,8 @@ pub(super) fn now_unix() -> i64 {
         .unwrap_or(0)
 }
 
-/// #3581: decide whether an abandoned `rebind_origin` inflight row is safe to
-/// reap. The predicate is a strict conjunction of "never-progressed,
-/// never-adopted, never-owned" signals so that a genuinely-live rebind
-/// (MonitorTriggered watcher rebind: `relay_owner_kind = Watcher`) or a row
-/// that has started relaying / been adopted is NEVER reaped:
-///
-///   * `rebind_origin` — only rebind-origin rows are in scope.
-///   * `turn_source == ExternalAdopted` — pins the predicate to the
-///     `recovery_engine` birth site; the `tmux` MonitorTriggered rebind
-///     (`turn_source = MonitorTriggered`, owner Watcher) is excluded twice.
-///   * `effective_relay_owner_kind() == None` — no live relay owner (also
-///     absorbs the legacy `watcher_owns_live_relay` bool).
-///   * `user_msg_id == 0 && current_msg_id == 0` — never adopted / no anchor.
-///   * `!terminal_delivery_committed` — not finalised.
-///   * `response_sent_offset == 0 && full_response.is_empty()` — nothing was
-///     ever streamed to Discord.
-///   * `last_offset == turn_start_offset` — the watcher never advanced past
-///     the birth offset (NOTE: a fresh rebind row is born with
-///     `last_offset == turn_start_offset == file_len`, which can be > 0 — the
-///     "no progress" test is offset equality, NOT `last_offset == 0`).
-///   * `restart_mode.is_none()` — planned restart / hot-swap rows own their
-///     own retention and are never reaped here.
-///
-/// When every structural conjunct holds, the row is reaped iff it is past its
-/// deadline OR it was born in a prior generation. `age_secs` is supplied by the
-/// caller (file-mtime age in the sweeper path) so legacy rows with no
-/// `rebind_origin_created_at_unix` stamp still age out via mtime.
-pub(super) fn should_reap_abandoned_rebind_origin(
-    state: &InflightTurnState,
-    age_secs: u64,
-    current_generation: u64,
-) -> bool {
-    if !state.rebind_origin {
-        return false;
-    }
-    let structurally_abandoned = state.turn_source == TurnSource::ExternalAdopted
-        && state.effective_relay_owner_kind() == RelayOwnerKind::None
-        && state.user_msg_id == 0
-        && state.current_msg_id == 0
-        && !state.terminal_delivery_committed
-        && state.response_sent_offset == 0
-        && state.full_response.is_empty()
-        && state.last_offset == state.turn_start_offset.unwrap_or(state.last_offset)
-        && state.restart_mode.is_none();
-    if !structurally_abandoned {
-        return false;
-    }
-
-    let deadline = state
-        .rebind_origin_deadline_secs
-        .unwrap_or_else(rebind_origin_deadline_secs_env);
-    let past_deadline = age_secs >= deadline;
-    let stale_generation = state
-        .rebind_origin_birth_generation
-        .is_some_and(|birth| birth != current_generation);
-    past_deadline || stale_generation
-}
-
-/// #3581: operator-visibility event for a reaped abandoned rebind-origin row
-/// (#3561 lifecycle stream). Mirrors the `evict_stale_generation` shape so the
-/// two boot-time reap reasons aggregate side by side.
-pub(super) fn emit_reap_abandoned_rebind_origin(
-    provider: &ProviderKind,
-    state: &InflightTurnState,
-    age_secs: u64,
-    current_generation: u64,
-    reason: &str,
-) {
-    crate::services::observability::emit_inflight_lifecycle_event(
-        provider.as_str(),
-        state.channel_id,
-        state.dispatch_id.as_deref(),
-        None,
-        None,
-        "reap_abandoned_rebind_origin",
-        serde_json::json!({
-            "reason": reason,
-            "age_secs": age_secs,
-            "deadline_secs": state
-                .rebind_origin_deadline_secs
-                .unwrap_or_else(rebind_origin_deadline_secs_env),
-            "birth_generation": state.rebind_origin_birth_generation,
-            "current_generation": current_generation,
-            "turn_source": state.turn_source.as_str(),
-            "tmux_session_name": state.tmux_session_name,
-        }),
-    );
-}
-
 pub(super) fn inflight_runtime_root() -> Option<PathBuf> {
     discord_inflight_root()
-}
-
-/// #3641: boot-time cleanup for orphan `discord_inflight/{provider}/*.json.lock`
-/// sidecars whose matching `.json` state file no longer exists. The lock file is
-/// never authoritative by itself: live turns are protected by the existence of
-/// the `.json` row, and fresh first-acquire locks are protected by the age floor.
-pub(super) fn reap_orphan_inflight_locks_in_root(root: &Path, now_unix: i64) -> usize {
-    let mut removed = 0usize;
-    for provider in [ProviderKind::Claude, ProviderKind::Codex] {
-        let provider_dir = inflight_provider_dir(root, &provider);
-        let provider_name = provider.as_str();
-        let Ok(lock_entries) = fs::read_dir(&provider_dir) else {
-            continue;
-        };
-
-        for lock_entry in lock_entries {
-            let Ok(lock_entry) = lock_entry else {
-                continue;
-            };
-            let lock_path = lock_entry.path();
-            if !is_inflight_json_lock_path(&lock_path) {
-                continue;
-            }
-            let Some(json_path) = inflight_json_path_for_lock(&lock_path) else {
-                continue;
-            };
-            if json_path.exists() {
-                continue;
-            }
-
-            let Ok(metadata) = lock_entry.metadata() else {
-                tracing::warn!(
-                    provider = provider_name,
-                    path = %lock_path.display(),
-                    "#3641 failed to stat inflight orphan-lock candidate"
-                );
-                continue;
-            };
-            if !metadata.is_file() {
-                continue;
-            }
-            let Some(mtime_unix) = metadata_mtime_unix_secs(&metadata) else {
-                continue;
-            };
-            let age_secs = now_unix.saturating_sub(mtime_unix).max(0);
-            if age_secs < ORPHAN_LOCK_REAP_MIN_AGE_SECS {
-                continue;
-            }
-            if json_path.exists() {
-                continue;
-            }
-
-            match fs::remove_file(&lock_path) {
-                Ok(()) => removed += 1,
-                Err(error) if error.kind() == std::io::ErrorKind::NotFound => {}
-                Err(error) => {
-                    tracing::warn!(
-                        provider = provider_name,
-                        path = %lock_path.display(),
-                        error = %error,
-                        "#3641 failed to remove orphan inflight lock file"
-                    );
-                }
-            }
-        }
-    }
-
-    if removed > 0 {
-        tracing::info!(
-            removed,
-            root = %root.display(),
-            "#3641 reaped orphan inflight lock file(s)"
-        );
-    }
-    removed
-}
-
-pub(super) fn reap_orphan_inflight_locks() -> usize {
-    let Some(root) = inflight_runtime_root() else {
-        return 0;
-    };
-    reap_orphan_inflight_locks_in_root(&root, now_unix())
 }
 
 /// #2235: expose the local `INFLIGHT_STATE_VERSION` so the recovery engine
@@ -423,322 +216,8 @@ pub(super) fn delete_inflight_state_file(provider: &ProviderKind, channel_id: u6
     fs::remove_file(path).is_ok()
 }
 
-/// #3581 (codex TOCTOU fix): outcome of a locked rebind-origin reap attempt so
-/// callers (and tests) can distinguish "reaped" from "skipped because the row
-/// was replaced/no-longer-eligible" from "the file was already gone".
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(super) enum RebindReapOutcome {
-    /// The row was re-validated under the lock and unlinked.
-    Reaped,
-    /// The on-disk row no longer satisfies the reap predicate (e.g. a live
-    /// intake/claim replaced it between the unlocked snapshot and the lock) —
-    /// deletion was intentionally skipped.
-    Skipped,
-    /// The state file was already absent (idempotent no-op) or unreadable.
-    Missing,
-    /// The advisory lock could not be acquired — the caller should retry later.
-    LockUnavailable,
-}
-
-/// #3581 (codex TOCTOU fix): true when the on-disk `locked` row is still the
-/// *same* abandoned rebind-origin orphan identified by the unlocked `snapshot`.
-///
-/// The placeholder sweeper (and the boot invalidator) pass an unlocked snapshot
-/// to `should_reap_abandoned_rebind_origin`; between that read and acquiring the
-/// sidecar lock a normal intake / TUI claim can persist a brand-new live
-/// inflight at the same `(provider, channel_id)` path. Re-running the reap
-/// predicate alone is not sufficient: a *new* rebind-origin orphan could be born
-/// (different birth stamp/generation) that also looks structurally abandoned.
-/// We therefore additionally require the row's birth identity to be unchanged
-/// (`rebind_origin_created_at_unix` + `rebind_origin_birth_generation`), so a
-/// replacement turn is never mistaken for the snapshotted orphan.
-fn rebind_row_identity_unchanged(snapshot: &InflightTurnState, locked: &InflightTurnState) -> bool {
-    locked.rebind_origin == snapshot.rebind_origin
-        && locked.rebind_origin_created_at_unix == snapshot.rebind_origin_created_at_unix
-        && locked.rebind_origin_birth_generation == snapshot.rebind_origin_birth_generation
-        && locked.turn_start_offset == snapshot.turn_start_offset
-}
-
-/// #3581 (codex TOCTOU fix): reap an abandoned rebind-origin orphan under the
-/// sidecar lock with a re-validate-then-unlink contract. This is the shared
-/// implementation behind both the periodic placeholder-sweeper path and the
-/// boot-time `invalidate_stale_generation` path so the two stay consistent.
-///
-/// Contract:
-///   1. Acquire the sidecar advisory lock for the row's path (the same
-///      non-reentrant `flock` every intake/claim/persist helper takes).
-///   2. **Reload** the current on-disk row under the lock.
-///   3. Confirm the reloaded row is still the *same* orphan as `snapshot`
-///      (`rebind_row_identity_unchanged`) AND still satisfies
-///      `should_reap_abandoned_rebind_origin` with a freshly recomputed age
-///      (created-at preferred, mtime fallback for legacy rows).
-///   4. Unlink **only** when both checks hold; otherwise skip (a live intake /
-///      claim replaced the orphan since the snapshot).
-///
-/// Returns the [`RebindReapOutcome`]; the caller emits observability on
-/// [`RebindReapOutcome::Reaped`].
-fn reap_abandoned_rebind_origin_locked_in_root(
-    root: &Path,
-    provider: &ProviderKind,
-    snapshot: &InflightTurnState,
-    current_generation: u64,
-) -> RebindReapOutcome {
-    let path = inflight_state_path(root, provider, snapshot.channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return RebindReapOutcome::LockUnavailable;
-    };
-    let Some(locked) = read_inflight_state_content(&path) else {
-        return RebindReapOutcome::Missing;
-    };
-    // The unlocked snapshot may have raced a replacement turn. Re-validate the
-    // current row's birth identity AND its eligibility under a freshly computed
-    // age before touching the file.
-    if !rebind_row_identity_unchanged(snapshot, &locked) {
-        return RebindReapOutcome::Skipped;
-    }
-    let age_secs = rebind_origin_age_secs(&path, &locked);
-    if !should_reap_abandoned_rebind_origin(&locked, age_secs, current_generation) {
-        return RebindReapOutcome::Skipped;
-    }
-    match fs::remove_file(&path) {
-        Ok(()) => RebindReapOutcome::Reaped,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => RebindReapOutcome::Missing,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = snapshot.channel_id,
-                error = %error,
-                "#3581 rebind reap remove_file failed under lock; treating as Missing"
-            );
-            RebindReapOutcome::Missing
-        }
-    }
-}
-
-/// #3581 (codex TOCTOU fix): env-rooted wrapper around
-/// [`reap_abandoned_rebind_origin_locked_in_root`] for the periodic
-/// placeholder-sweeper path. Returns `true` iff the row was re-validated under
-/// the lock and unlinked (so the sweeper only counts/emits genuine reaps).
-pub(super) fn reap_abandoned_rebind_origin_locked(
-    provider: &ProviderKind,
-    snapshot: &InflightTurnState,
-    current_generation: u64,
-) -> bool {
-    let Some(root) = inflight_runtime_root() else {
-        return false;
-    };
-    reap_abandoned_rebind_origin_locked_in_root(&root, provider, snapshot, current_generation)
-        == RebindReapOutcome::Reaped
-}
-
-/// #3635: locked re-validate-then-unlink for the *dead-watcher* rebind-origin
-/// reap path. The expensive tmux liveness probe runs OUTSIDE this lock (see
-/// [`sweep_reap_dead_watcher_rebind_origin`], off the async runtime via
-/// `spawn_blocking`); under the lock we re-check only CHEAP, fs-only conditions —
-/// no tmux subprocess — so the per-channel sidecar lock is never held across
-/// blocking subprocess I/O (codex review ISSUE 2).
-///
-/// What the lock DOES close: the row-replacement race — a replacement turn that
-/// rewrote this sidecar path is rejected by the birth-identity guard, and a row
-/// that progressed (structural conjunction) is rejected. What it does NOT fully
-/// close (codex review ISSUE 3): a watcher whose tmux *pane* re-appears between
-/// the unlocked probe and this unlink. That window is tiny and the consequence
-/// is benign — the reaped artifact is a zero-msg-id orphan that a genuinely
-/// re-adopting watcher simply re-creates. As a last cheap guard we re-read the
-/// runtime-activity mtime under the lock, so a watcher that resumed *writing*
-/// (the most common resurrection signal) is still observed and the reap Skips.
-fn reap_dead_watcher_rebind_origin_locked_in_root(
-    root: &Path,
-    provider: &ProviderKind,
-    snapshot: &InflightTurnState,
-    current_generation: u64,
-) -> RebindReapOutcome {
-    let path = inflight_state_path(root, provider, snapshot.channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return RebindReapOutcome::LockUnavailable;
-    };
-    let Some(locked) = read_inflight_state_content(&path) else {
-        return RebindReapOutcome::Missing;
-    };
-    // Birth-identity guard (same as the None-owner path): a replacement turn at
-    // the same sidecar path must never be mistaken for the snapshotted orphan.
-    if !rebind_row_identity_unchanged(snapshot, &locked) {
-        return RebindReapOutcome::Skipped;
-    }
-    let age_secs = rebind_origin_age_secs(&path, &locked);
-    // Re-check the cheap, fs-only half of the predicate under the lock (structure
-    // + deadline/generation). The tmux liveness probe already ran unlocked; we do
-    // NOT re-spawn it here (ISSUE 2 — no subprocess under the sidecar lock).
-    if !dead_watcher_rebind_structurally_reapable(&locked, age_secs, current_generation) {
-        return RebindReapOutcome::Skipped;
-    }
-    // Last cheap guard: a watcher that resumed writing its jsonl/`.generation`
-    // since the unlocked probe reads as recent activity ⇒ alive ⇒ Skip. Pure fs
-    // stat (no subprocess), safe under the lock.
-    if let Some(session) = locked.tmux_session_name.as_deref() {
-        let session = session.trim();
-        if !session.is_empty() && watcher_runtime_activity_recent(session) {
-            return RebindReapOutcome::Skipped;
-        }
-    }
-    match fs::remove_file(&path) {
-        Ok(()) => RebindReapOutcome::Reaped,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => RebindReapOutcome::Missing,
-        Err(error) => {
-            tracing::warn!(
-                provider = %provider.as_str(),
-                channel = snapshot.channel_id,
-                error = %error,
-                "#3635 dead-watcher rebind reap remove_file failed under lock; treating as Missing"
-            );
-            RebindReapOutcome::Missing
-        }
-    }
-}
-
-/// #3635: env-rooted wrapper around
-/// [`reap_dead_watcher_rebind_origin_locked_in_root`] for the periodic
-/// placeholder-sweeper path. Returns `true` iff the row was re-validated under
-/// the lock (cheap fs-only re-checks; no tmux subprocess) and unlinked.
-pub(super) fn reap_dead_watcher_rebind_origin_locked(
-    provider: &ProviderKind,
-    snapshot: &InflightTurnState,
-    current_generation: u64,
-) -> bool {
-    let Some(root) = inflight_runtime_root() else {
-        return false;
-    };
-    reap_dead_watcher_rebind_origin_locked_in_root(&root, provider, snapshot, current_generation)
-        == RebindReapOutcome::Reaped
-}
-
-/// #3635: the placeholder sweeper's single-call entry point for the dead-watcher
-/// rebind-origin reap. Returns `true` only when the row was genuinely unlinked.
-/// Three stages, ordered cheapest-first:
-///   1. fs-only structural gate ([`dead_watcher_rebind_structurally_reapable`]) —
-///      most rows fail here with no tmux probe;
-///   2. the production [`RuntimeWatcherLiveness`] proven-dead probe, run on a
-///      `spawn_blocking` thread (it spawns tmux subprocesses) so the async
-///      sweeper runtime is never blocked, and OUTSIDE any lock (codex review
-///      ISSUE 2);
-///   3. the locked re-validate ([`reap_dead_watcher_rebind_origin_locked`]),
-///      which re-checks only cheap fs-only conditions under the sidecar lock.
-///
-/// Deliberately NOT called from the boot path
-/// (`invalidate_stale_generation_in_root`): at cold start a just-restarted live
-/// watcher's session reads as dead, so the liveness gate only ever fires in the
-/// warm 30s sweeper where a real runtime probe is meaningful (#3154/#3540 / the
-/// keystone boot-preservation invariant).
-pub(super) async fn sweep_reap_dead_watcher_rebind_origin(
-    provider: &ProviderKind,
-    state: &InflightTurnState,
-    age_secs: u64,
-    current_generation: u64,
-) -> bool {
-    // (1) Cheap, fs-only structural gate first — skip the subprocess entirely for
-    // the common case.
-    if !dead_watcher_rebind_structurally_reapable(state, age_secs, current_generation) {
-        return false;
-    }
-    // (2) The proven-dead probe spawns tmux subprocesses; run it off the async
-    // runtime via `spawn_blocking`, outside any lock. A join failure (panic /
-    // runtime shutdown) is treated as "unknown ⇒ preserve".
-    let probe_state = state.clone();
-    let proven_dead =
-        tokio::task::spawn_blocking(move || RuntimeWatcherLiveness.is_proven_dead(&probe_state))
-            .await
-            .unwrap_or(false);
-    if !proven_dead {
-        return false;
-    }
-    // (3) Locked re-validate (cheap fs-only re-checks; no subprocess under lock).
-    reap_dead_watcher_rebind_origin_locked(provider, state, current_generation)
-}
-
 fn now_string() -> String {
     chrono::Local::now().format("%Y-%m-%d %H:%M:%S").to_string()
-}
-
-/// Parse the persisted `started_at` (`now_string` localtime form) back into
-/// a Unix timestamp. Returns `None` for unparseable values so callers can
-/// fall back to a wall-clock derived approximation.
-pub(super) fn parse_started_at_unix(started_at: &str) -> Option<i64> {
-    let naive = chrono::NaiveDateTime::parse_from_str(started_at, "%Y-%m-%d %H:%M:%S").ok()?;
-    chrono::Local
-        .from_local_datetime(&naive)
-        .single()
-        .map(|local| local.timestamp())
-}
-
-/// Parse a persisted `updated_at` field (same `now_string` localtime form
-/// as `started_at`) back into a Unix timestamp. Wrapper kept distinct from
-/// `parse_started_at_unix` purely for call-site readability — both fields
-/// share the same encoding but represent different lifecycle moments.
-pub(super) fn parse_updated_at_unix(updated_at: &str) -> Option<i64> {
-    parse_started_at_unix(updated_at)
-}
-
-/// #1446 stall-deadlock recovery: returns `true` when the persisted
-/// `updated_at` of an inflight state is older than
-/// `threshold_secs` seconds relative to `now_unix_secs`.
-///
-/// Returns `false` if `updated_at` is unparseable — staleness should never
-/// be inferred from missing data. This keeps the helper safe to call from
-/// the THREAD-GUARD and stall-watchdog paths even when a partially
-/// migrated state file is on disk.
-pub(super) fn inflight_state_is_stale(
-    state: &InflightTurnState,
-    now_unix_secs: i64,
-    threshold_secs: u64,
-) -> bool {
-    let Some(updated_at_unix) = parse_updated_at_unix(&state.updated_at) else {
-        return false;
-    };
-    let age_secs = now_unix_secs.saturating_sub(updated_at_unix);
-    age_secs >= 0 && (age_secs as u64) >= threshold_secs
-}
-
-fn inflight_state_started_at_is_stale(
-    state: &InflightTurnState,
-    now_unix_secs: i64,
-    threshold_secs: u64,
-) -> bool {
-    let Some(started_at_unix) = parse_started_at_unix(&state.started_at) else {
-        return false;
-    };
-    let age_secs = now_unix_secs.saturating_sub(started_at_unix);
-    age_secs >= 0 && (age_secs as u64) >= threshold_secs
-}
-
-/// A TUI-direct `ExternalInput` row can be born as a bridge-owned synthetic
-/// claim before the bridge tail has created the real response placeholder. If
-/// dcserver restarts in that narrow window, the tail is gone and the row has no
-/// live relay owner left. Treat that shape as stale after the normal inflight
-/// threshold so prompt scanners / idle relays / health recovery do not block
-/// forever on a row that cannot make progress.
-pub(super) fn ownerless_external_input_inflight_is_stale_at(
-    state: &InflightTurnState,
-    now_unix_secs: i64,
-) -> bool {
-    state.turn_source == TurnSource::ExternalInput
-        && state.effective_relay_owner_kind() == RelayOwnerKind::None
-        && state.injected_prompt_message_id.is_some()
-        && state.current_msg_id == 0
-        && state.response_sent_offset == 0
-        && state.full_response.trim().is_empty()
-        && state.last_watcher_relayed_offset.is_none()
-        && !state.terminal_delivery_committed
-        && (inflight_state_is_stale(state, now_unix_secs, INFLIGHT_STALENESS_THRESHOLD_SECS)
-            || (state.restart_mode.is_some()
-                && inflight_state_started_at_is_stale(
-                    state,
-                    now_unix_secs,
-                    INFLIGHT_STALENESS_THRESHOLD_SECS,
-                )))
-}
-
-pub(super) fn ownerless_external_input_inflight_is_stale(state: &InflightTurnState) -> bool {
-    ownerless_external_input_inflight_is_stale_at(state, now_unix())
 }
 
 fn turn_id_for_state(state: &InflightTurnState) -> Option<String> {
@@ -1381,287 +860,6 @@ pub(super) fn save_inflight_state_if_matches_identity_in_root(
             );
             GuardedSaveOutcome::IoError
         }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// #3077: typed status-panel ownership writes.
-//
-// `status_message_id` is the de-facto "this turn owns status panel M" pointer,
-// and several independent actors (turn-bridge completion fallback, tmux watcher
-// TUI-direct publish/orphan-cleanup, placeholder sweeper) used to mutate it via
-// a non-atomic `load_inflight_state(...)` → `state.status_message_id = …` →
-// `save_inflight_state(...)` triple. Because the read and the write were not
-// serialized against each other, a newer turn that rebound the panel between a
-// stale actor's load and its blind `= None` could have its panel silently
-// orphaned (the #3099/#3100/#3105/#3107 panel-lifecycle race family).
-//
-// These helpers centralize that read-modify-write behind intentful operations
-// that hold the same `lock_inflight_state_path` sidecar flock across the whole
-// compare-and-set, exactly like `save_inflight_state_if_absent`. Callers no
-// longer touch the field directly; they declare *what* they own and *under
-// which precondition*, and the store enforces it atomically.
-
-/// Per-turn precondition for `bind_status_panel`. Lets each caller carry its
-/// own ownership invariant into the lock-held read-modify-write so the guard
-/// check and the write cannot be split by a concurrent writer (TOCTOU).
-#[derive(Debug, Clone, Default)]
-pub(in crate::services::discord) struct StatusPanelBindGuard {
-    /// Bind only when the on-disk row still belongs to this `user_msg_id`.
-    /// `None` means "do not guard on user_msg_id" (used by callers that have
-    /// already established identity another way). Mirrors the turn-bridge
-    /// status-panel-v2 completion fallback guard.
-    pub require_user_msg_id: Option<u64>,
-    /// Bind only when the on-disk row still matches this full turn identity
-    /// (user_msg_id + started_at + tmux_session_name). Mirrors the tmux
-    /// watcher TUI-direct publish guard that defeats turn handoff during the
-    /// Discord send await.
-    pub require_identity: Option<InflightTurnIdentity>,
-    /// When true, do not overwrite a real (non-synthetic) panel id already on
-    /// the row — an overlapping actor already published the canonical panel
-    /// and our send is a duplicate (#3003 reclaim path). Synthetic-headless
-    /// ids do not count as "already set".
-    pub skip_if_panel_already_set: bool,
-}
-
-/// Outcome of a `bind_status_panel` attempt.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(in crate::services::discord) enum StatusPanelBindOutcome {
-    /// The row was found, passed the guard, and now carries `msg_id`.
-    Bound,
-    /// The row already carried `msg_id`; nothing was written.
-    AlreadyBound,
-    /// The row exists but a DIFFERENT real panel id is already set and
-    /// `skip_if_panel_already_set` was requested — left untouched. Carries the
-    /// row's currently-owned (real) panel id as observed under the same flock,
-    /// so the caller can adopt the row's actual panel without a second
-    /// (racy) re-read of the inflight row.
-    SkippedPanelAlreadySet(u64),
-    /// No inflight row exists for `(provider, channel_id)`.
-    Missing,
-    /// The on-disk row did not satisfy `require_user_msg_id` /
-    /// `require_identity` — left untouched (a different turn now owns the row).
-    GuardMismatch,
-    /// Filesystem / serialization failure while persisting the bind.
-    IoError,
-}
-
-/// Intentful "this turn now owns status panel `msg_id`" write. Performs the
-/// guard check and the field set atomically under the inflight sidecar flock,
-/// so it is consistent with `save_inflight_state` / `save_inflight_state_if_absent`.
-pub(in crate::services::discord) fn bind_status_panel(
-    provider: &ProviderKind,
-    channel_id: u64,
-    msg_id: u64,
-    guard: &StatusPanelBindGuard,
-) -> StatusPanelBindOutcome {
-    let Some(root) = inflight_runtime_root() else {
-        return StatusPanelBindOutcome::IoError;
-    };
-    bind_status_panel_in_root(&root, provider, channel_id, msg_id, guard)
-}
-
-fn bind_status_panel_in_root(
-    root: &Path,
-    provider: &ProviderKind,
-    channel_id: u64,
-    msg_id: u64,
-    guard: &StatusPanelBindGuard,
-) -> StatusPanelBindOutcome {
-    let path = inflight_state_path(root, provider, channel_id);
-    if let Some(parent) = path.parent()
-        && fs::create_dir_all(parent).is_err()
-    {
-        return StatusPanelBindOutcome::IoError;
-    }
-    // Hold the sidecar flock across load → guard → set so a concurrent
-    // writer cannot land between the guard check and the write.
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return StatusPanelBindOutcome::IoError;
-    };
-    let Some(mut state) = load_inflight_state_unlocked(&path) else {
-        return StatusPanelBindOutcome::Missing;
-    };
-    if let Some(expected) = guard.require_user_msg_id
-        && state.user_msg_id != expected
-    {
-        return StatusPanelBindOutcome::GuardMismatch;
-    }
-    if let Some(expected) = guard.require_identity.as_ref()
-        && !expected.matches_state(&state)
-    {
-        return StatusPanelBindOutcome::GuardMismatch;
-    }
-    // Same-id re-bind is idempotent and must classify as `AlreadyBound`
-    // REGARDLESS of `skip_if_panel_already_set` — an idempotent re-bind of the
-    // panel this row already owns is a no-op, not a "duplicate skip". Checking
-    // the skip flag first (#3077 codex P2 #1) misclassified a same-id re-bind as
-    // `SkippedPanelAlreadySet`, which the TUI-direct caller then routed to a
-    // DELETE of the row's own already-bound panel. Order: same-id → AlreadyBound;
-    // else a DIFFERENT real id already set + skip flag → SkippedPanelAlreadySet.
-    if state.status_message_id == Some(msg_id) {
-        return StatusPanelBindOutcome::AlreadyBound;
-    }
-    if guard.skip_if_panel_already_set && status_panel_id_is_real(state.status_message_id) {
-        // Safe: `status_panel_id_is_real` only returns true for `Some(real)`,
-        // and we already handled `Some(msg_id)` above, so this is a DIFFERENT
-        // real panel id. Carry it so the caller adopts the row's owned panel.
-        return StatusPanelBindOutcome::SkippedPanelAlreadySet(
-            state.status_message_id.unwrap_or_default(),
-        );
-    }
-    state.status_message_id = Some(msg_id);
-    match persist_under_lock(
-        root,
-        &path,
-        &state,
-        "src/services/discord/inflight.rs:bind_status_panel_in_root",
-    ) {
-        Ok(()) => StatusPanelBindOutcome::Bound,
-        Err(_) => StatusPanelBindOutcome::IoError,
-    }
-}
-
-/// Per-turn precondition for `clear_status_panel_if_current`. The msg-id
-/// compare-and-clear is unconditional; these add the caller's extra ownership
-/// guards (so a sweeper/cleanup that loaded a stale snapshot does not clear a
-/// row a newer turn already advanced).
-#[derive(Debug, Clone, Default)]
-pub(in crate::services::discord) struct StatusPanelClearGuard {
-    /// Clear only when the on-disk row still belongs to this `user_msg_id`.
-    pub require_user_msg_id: Option<u64>,
-    /// Clear only when the on-disk row still carries this `current_msg_id`
-    /// (placeholder sweeper convergence guard).
-    pub require_current_msg_id: Option<u64>,
-    /// Clear only when the on-disk row's tmux session still matches (watcher
-    /// orphan-cleanup guard).
-    pub require_tmux_session_name: Option<String>,
-}
-
-/// Compare-and-clear: clears `status_message_id` ONLY when it currently equals
-/// `msg_id` (and every guard precondition holds). Returns `true` iff it
-/// cleared. This is the #3077 hardening — a blind `= None` becomes a
-/// compare-and-clear so a panel a *newer* turn rebound is never wiped by a
-/// stale actor that loaded an older snapshot.
-pub(in crate::services::discord) fn clear_status_panel_if_current(
-    provider: &ProviderKind,
-    channel_id: u64,
-    msg_id: u64,
-    guard: &StatusPanelClearGuard,
-) -> bool {
-    let Some(root) = inflight_runtime_root() else {
-        return false;
-    };
-    clear_status_panel_if_current_in_root(&root, provider, channel_id, msg_id, guard)
-}
-
-fn clear_status_panel_if_current_in_root(
-    root: &Path,
-    provider: &ProviderKind,
-    channel_id: u64,
-    msg_id: u64,
-    guard: &StatusPanelClearGuard,
-) -> bool {
-    let path = inflight_state_path(root, provider, channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return false;
-    };
-    let Some(mut state) = load_inflight_state_unlocked(&path) else {
-        return false;
-    };
-    if state.status_message_id != Some(msg_id) {
-        return false;
-    }
-    if let Some(expected) = guard.require_user_msg_id
-        && state.user_msg_id != expected
-    {
-        return false;
-    }
-    if let Some(expected) = guard.require_current_msg_id
-        && state.current_msg_id != expected
-    {
-        return false;
-    }
-    if let Some(expected) = guard.require_tmux_session_name.as_deref()
-        && state.tmux_session_name.as_deref() != Some(expected)
-    {
-        return false;
-    }
-    state.status_message_id = None;
-    persist_under_lock(
-        root,
-        &path,
-        &state,
-        "src/services/discord/inflight.rs:clear_status_panel_if_current_in_root",
-    )
-    .is_ok()
-}
-
-/// #3351: compare-and-clear for the persisted relay-placeholder id, mirroring
-/// `clear_status_panel_if_current` (#3077): exact `msg_id` only, placeholderless
-/// turns (`current_msg_id == user_msg_id`) untouched, optional tmux-session guard.
-pub(in crate::services::discord) fn clear_current_msg_if_matches(
-    provider: &ProviderKind,
-    channel_id: u64,
-    msg_id: u64,
-    require_tmux_session_name: Option<&str>,
-) -> bool {
-    let Some(root) = inflight_runtime_root() else {
-        return false;
-    };
-    clear_current_msg_if_matches_in_root(
-        &root,
-        provider,
-        channel_id,
-        msg_id,
-        require_tmux_session_name,
-    )
-}
-
-fn clear_current_msg_if_matches_in_root(
-    root: &Path,
-    provider: &ProviderKind,
-    channel_id: u64,
-    msg_id: u64,
-    require_tmux_session_name: Option<&str>,
-) -> bool {
-    let path = inflight_state_path(root, provider, channel_id);
-    let Ok(_lock) = lock_inflight_state_path(&path) else {
-        return false;
-    };
-    let Some(mut state) = load_inflight_state_unlocked(&path) else {
-        return false;
-    };
-    if msg_id == 0 || state.current_msg_id != msg_id {
-        // A newer turn already advanced the anchor — never wipe it.
-        return false;
-    }
-    if state.user_msg_id != 0 && state.current_msg_id == state.user_msg_id {
-        // Placeholderless turn: anchor mirrors the user's own message id.
-        return false;
-    }
-    if let Some(expected) = require_tmux_session_name
-        && state.tmux_session_name.as_deref() != Some(expected)
-    {
-        return false;
-    }
-    state.current_msg_id = 0;
-    persist_under_lock(
-        root,
-        &path,
-        &state,
-        "src/services/discord/inflight.rs:clear_current_msg_if_matches_in_root",
-    )
-    .is_ok()
-}
-
-/// `true` when `id` is a real Discord panel id (present and not a synthetic
-/// headless placeholder). Mirrors `turn_bridge::normalize_status_panel_message_id`
-/// without pulling in the serenity `MessageId` newtype.
-fn status_panel_id_is_real(id: Option<u64>) -> bool {
-    match id {
-        Some(value) => !super::is_synthetic_headless_message_id_raw(value),
-        None => false,
     }
 }
